Extract request and response classes to helper package
