+++ src/main/java/simplefixer/controller/MinifixController.java
@@ -12,131 +12,10 @@ import simplefixer.model.IssueType;
 import simplefixer.model.Minifix;
 import simplefixer.repository.IssueTypeRepository;
 import simplefixer.repository.MinifixRepository;
+import simplefixer.helper.*;
 
 import java.io.IOException;
 import java.util.*;
-
-// Serializer class for post response
-class PostResponse {
-    private Integer statusCode;
-    private String statusMessage;
-    private Integer fixId;
-
-    PostResponse(Integer statusCode, String statusMessage) {
-        this.statusCode = statusCode;
-        this.statusMessage = statusMessage;
-        this.fixId = -1;
-    }
-
-    PostResponse(Integer statusCode, String statusMessage, Integer fixId) {
-        this.statusCode = statusCode;
-        this.statusMessage = statusMessage;
-        this.fixId = fixId;
-    }
-}
-
-class IssueResponse {
-    private Integer id;
-    private Integer issueTypeId;
-    private String issueTypeDesc;
-    private String fileName;
-    private Integer lineNumber;
-    private Integer columnNumber;
-    private Boolean isFixed;
-
-    IssueResponse(Integer id, Integer issueTypeId, String issueTypeDesc,String fileName, Integer lineNumber,Integer columnNumber, Boolean isFixed) {
-        this.id = id;
-        this.issueTypeId = issueTypeId;
-        this.issueTypeDesc = issueTypeDesc;
-        this.fileName = fileName;
-        this.lineNumber = lineNumber;
-        this.columnNumber = columnNumber;
-        this.isFixed = isFixed;
-    }
-}
-
-// Serializer class for get response
-class GetResponse {
-    private Integer statusCode;
-    private String statusMessage;
-    private Integer fixId;
-    private String s3Link;
-    List<IssueResponse> issueResponses;
-
-    private List<IssueResponse> buildIssueResponses(List<Minifix> minifixes) {
-        List<IssueResponse> issueResponses = new ArrayList<>();
-        for (Minifix minifix : minifixes) {
-            issueResponses.add(new IssueResponse(
-                minifix.getId(),
-                minifix.getIssueType().getId(),
-                minifix.getIssueType().getIssueDescription(),
-                minifix.getFileName(),
-                minifix.getLineNumber(),
-                minifix.getColumnNumber(),
-                minifix.getIsFixed())
-            );
-        }
-        return issueResponses;
-    }
-
-    GetResponse(Integer statusCode, String statusMessage) {
-        this.statusCode = statusCode;
-        this.statusMessage = statusMessage;
-    }
-
-    GetResponse(Integer statusCode, String statusMessage, Integer fixId, String s3Link, List<Minifix> minifixes) {
-        this.statusCode = statusCode;
-        this.statusMessage = statusMessage;
-        this.fixId = fixId;
-        this.s3Link = s3Link;
-        this.issueResponses = buildIssueResponses(minifixes);
-    }
-}
-
-class CodeGenPostRequest {
-    private String url;
-    private String username;
-    private String password;
-    private Set<Integer> fileIds;
-
-    CodeGenPostRequest(String url, String username,String password, Set<Integer> fileIds) {
-        this.url = url;
-        this.username = username;
-        this.password = password;
-        this.fileIds = fileIds;
-    }
-}
-
-class CodeGenPostResponse {
-    private Integer id;
-
-    CodeGenPostResponse(Integer id) {
-        this.id = id;
-    }
-
-    public Integer getId() {
-        return id;
-    }
-}
-
-// Serializer class for codegen response
-class CodeGenGetResponse {
-    private String status;
-    private String Url;
-
-    CodeGenGetResponse(String status, String s3Link) {
-        this.status = status;
-        this.Url = s3Link;
-    }
-
-    public String getStatus() {
-        return status;
-    }
-
-    public String getUrl() {
-        return Url;
-    }
-}
 
 @RestController
 public class MinifixController {
