+++ src/main/java/simplefixer/helper/CodeGenPostResponse.java
@@ -1,0 +1,13 @@
+package simplefixer.helper;
+
+public class CodeGenPostResponse {
+    private Integer id;
+
+    public CodeGenPostResponse(Integer id) {
+        this.id = id;
+    }
+
+    public Integer getId() {
+        return id;
+    }
+}
